+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - nth function
 - has function
 - prop function
+- groupBy function
 
 ### Changed
 - added more types